test/distributed/dqn/server/train_node/requirements.txt
numpy==1.22.4
redis==3.5.3
gym==0.21.0
torch==1.11.0
pyyaml==6.0
